+++ src/signin/Signin.tsx
@@ -1,11 +1,42 @@
 import React, {useState} from 'react';
-import {Button, Text, TextInput, SafeAreaView} from 'react-native';
-import {validate} from '../utils/validation';
+import {
+  TouchableOpacity,
+  ActivityIndicator,
+  Text,
+  TextInput,
+  SafeAreaView,
+  Alert,
+} from 'react-native';
+import {validateEmail, validatePassword} from '../utils/validation';
 import {NavigationComponentProps} from 'react-native-navigation';
+import {goToHome} from '../utils/navigation';
+import {loginRequest} from '../utils/apollo';
 
 export const LoginScreen = (props: NavigationComponentProps) => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [loading, setLoading] = useState(false);
+
+  const validate = async () => {
+    const emailValidation = validateEmail(email);
+    const passwordValidation = validatePassword(password);
+    try {
+      if (emailValidation !== '') {
+        return Alert.alert(emailValidation);
+      }
+
+      if (validatePassword(password) !== '') {
+        return Alert.alert(passwordValidation);
+      }
+      setLoading(true);
+      await loginRequest(email, password);
+      setLoading(false);
+      goToHome(props.componentId);
+    } catch {
+      setLoading(false);
+      throw 'Não foi possivel realizar o login';
+    }
+  };
 
   return (
     <SafeAreaView>
@@ -23,10 +54,13 @@ export const LoginScreen = (props: NavigationComponentProps) => {
         secureTextEntry={true}
       />
 
-      <Button
-        title="Entrar"
-        onPress={() => validate(email, password, props.componentId)}
-      />
+      {!loading ? (
+        <TouchableOpacity onPress={() => validate()}>
+          <Text>Entrar</Text>
+        </TouchableOpacity>
+      ) : (
+        <ActivityIndicator />
+      )}
     </SafeAreaView>
   );
 };
