+++ src/utils/validation.ts
@@ -1,26 +1,5 @@
-import {Alert} from 'react-native';
-import {loginRequest} from './apollo';
-
 const regexEmail = /.+[@].+\.com/;
 const regexPassword = /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{7,}$/;
-
-export const validate = (
-  email: string,
-  password: string,
-  componentId: string,
-) => {
-  const emailValidation = validateEmail(email);
-  const passwordValidation = validatePassword(password);
-  if (emailValidation !== '') {
-    return Alert.alert(emailValidation);
-  }
-
-  if (validatePassword(password) !== '') {
-    return Alert.alert(passwordValidation);
-  }
-
-  loginRequest(email, password, componentId);
-};
 
 export const validateEmail = (email: string) => {
   if (email === '') {
